+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - cli: Add delay command to update todo due with datetime, time, or tomorrow keyword
+- core: Set default time range for event if not specified
 
 ### Fixed